#438 refactor: CR round 3 — split navigation into browser.py, document its boundary
Navigation resilience had nothing to do with axe, and test_browser_smoke.py
imported the axe SHA verification just to navigate. goto_with_retry and its
predicate move to tests/api/admin/browser.py; axe.py keeps the axe plumbing.

Also record the retry's scope boundary (goto only, not mid-test crashes) and
that net::ERR_ABORTED is the loosest signature, in both the module docstring
and docs/TESTING.md.
